+++ vector/src/debug/java/im/vector/riotx/features/debug/sas/SasEmojiItem.kt
@@ -25,7 +25,7 @@ import im.vector.riotx.core.epoxy.VectorEpoxyHolder
 import im.vector.riotx.core.epoxy.VectorEpoxyModel
 
 @EpoxyModelClass(layout = im.vector.riotx.R.layout.item_sas_emoji)
-abstract class ItemSasEmoji : VectorEpoxyModel<ItemSasEmoji.Holder>() {
+abstract class SasEmojiItem : VectorEpoxyModel<SasEmojiItem.Holder>() {
 
     @EpoxyAttribute
     var index: Int = 0
