+++ vector/src/main/java/im/vector/riotx/core/epoxy/bottomsheet/BottomSheetSendStateItem.kt
@@ -1,0 +1,53 @@
+/*
+ * Copyright 2019 New Vector Ltd
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+package im.vector.riotx.core.epoxy.bottomsheet
+
+import android.view.View
+import android.widget.TextView
+import androidx.annotation.DrawableRes
+import androidx.core.view.isVisible
+import com.airbnb.epoxy.EpoxyAttribute
+import com.airbnb.epoxy.EpoxyModelClass
+import im.vector.riotx.R
+import im.vector.riotx.core.epoxy.VectorEpoxyHolder
+import im.vector.riotx.core.epoxy.VectorEpoxyModel
+
+/**
+ * A send state for bottom sheet.
+ */
+@EpoxyModelClass(layout = R.layout.item_bottom_sheet_message_status)
+abstract class BottomSheetSendStateItem : VectorEpoxyModel<BottomSheetSendStateItem.Holder>() {
+
+    @EpoxyAttribute
+    var showProgress: Boolean = false
+    @EpoxyAttribute
+    lateinit var text: CharSequence
+    @EpoxyAttribute
+    @DrawableRes
+    var drawableStart: Int = 0
+
+    override fun bind(holder: Holder) {
+        holder.progress.isVisible = showProgress
+        holder.text.setCompoundDrawablesWithIntrinsicBounds(drawableStart, 0, 0, 0)
+        holder.text.text = text
+    }
+
+    class Holder : VectorEpoxyHolder() {
+        val progress by bind<View>(R.id.messageStatusProgress)
+        val text by bind<TextView>(R.id.messageStatusText)
+    }
+}
